Refactor
Create src/process.js and move main code into functions, for easier reuse and testing.

Change SOLR_MAX_ROWS to options["RowsPerPage"], so it is consistent with other options
and so it can be changed easily later.
Also increase the default from 100 to 1000, the most IATI Datastore API allows.

Change loadJsonFile to loadFiltersFileData so that the date object building is in a util function
for easier reuse and testing.